fix: update ascii symbols for angled gate adjoint
